Fix bug with scopeFrameToken
The API contract of scopeFrameToken is supposed to be that it returns an identifier that remains the same for components that share the same logical scope frame across updates.

It was never actually implemented this way; oops. Instead, it returned the scope frame directly, which was unique in each version of the tree but not the same across tree updates.

Now that component scope frames have global identifiers, we can use that.