Revert usage of slather
This dependency on building and testing ComponentKit has blocked several engineers from contributing over the past couple of months. Code coverage is a good thing, and we want to have it, but the reality is that this is too fragile as integrated. Until we have a better solution, we should just revert this to unblock casual contributors.

Revert "README: Add Coveralls badge"

This reverts commit e1a60212eefd047f3252033d6236f0c7ca0255de.

Revert "Generate coverage stats using slather/Coveralls"

This reverts commit d942965412ad1080292f3b72dcde6347283b8bd6.