Fix pcomponents for CKTransactionalComponentDataSource
pcomponents relies on being able to find the layout of the component
hierarchy from the root view. The mapping existed for the legacy data
source but doesn't exist for the transactional one.

This adds a mapping from root view to layout in the transactional data
source that can be accessed using an internal header. It also fixes up
pcomponents so that pcomponents uses this mapping.