Fix stateful view reuse pool bug with superview changes
Credit to @RomanoffAlex for finding this and proposing a fix. His description of the problem follows:

If we have preferredSuperviewMap like the following:

```
superview: [view].
```

and allViews:

```
[view].
```

Imagine that we call `viewWithPreferredSuperview(UIView *preferredSuperview)` with some `preferredSuperview` that is not in the map and a corresponding stateful view (which was picked to be recycled) doesn't have a superview.

In this case we will remove the `preferredSuperview` from `allViews` but keep it in `preferredSuperviewMap` which is really bad because if we later call `viewWithPreferredSuperview` with a superview that is a key in the map, we will reuse the view that already is used by some component that means we are stealing a stateful view from another superview that is on screen.

To fix this, get rid of the complexity of `preferredSuperviewMap` entirely. The map does not grow so large that its algorithmic efficiency is needed. Replace it with a simple linear search.