Consider supplementary views in collection view bounds animation
This extends the component bounds animation so it considers the headers
or footers of the collection view as well. I changed the existing
implementation to use `layoutAttributesForElementsInRect:` rather than
`indexPathsForVisibleItems` and `layoutAttributesForItemAtIndexPath:`.
This makes it a little easier to retrieve layout attributes for
offscreen cells (you just extend the rect) and also works for
supplementary views without resorting to iOS9 specific APIs like
`visibleSupplementaryViewsOfKind:`.