Don't dispatch the deleter off the main thread if the layout vector is empty
When you set an CKComponentLayout ivar for the first time, it’s already
initialized with the default initializer. Because of this, the deleter
is triggered with an empty array. The innocuous setting of an ivar
therefore triggers a semi-expensive dispatch. It’s a lot cheaper to
simply delete this object on the main thread.