[ck] Fix controller.component = nil before mount
As discussed in https://github.com/facebook/componentkit/issues/504
CKComponentController component can be nil until first mount. This diff
changes the initializer of CKComponentController to instead take in a
component on init, and changes the initialization order for controllers
inside scope handles.

[ck] Fix controller.component = nil before mount

As discussed in https://github.com/facebook/componentkit/issues/504
CKComponentController component can be nil until first mount. This diff
changes the initializer of CKComponentController to instead take in a
component on init, and changes the initialization order for controllers
inside scope handles.