Refactor CKComponentActionAttribute
Summary:
When this function was originally designed, it was a safe assumption that the number of unique `CKComponentAction` identifiers in the app would be at most ~200. Thus it seemed reasonable to allocate a separate "forwarder" for each identifier and keep the corresponding object in the map forever.

With the introduction of scope-based and block-based actions, this assumption is dangerous; there could potentially be tens of thousands of distinct identifiers, and each distinct action will result in the allocation of a separate forwarder that is essentially leaked forever.

Refactor by keying forwarders on `UIControlEvents` instead of `CKComponentAction::identifier`. This should result in at most ~16 forwarders being active at any given time, regardless of the number of actions.
Closes https://github.com/facebook/componentkit/pull/848

Reviewed By: dshahidehpour, ocrickard

Differential Revision: D5563932

Pulled By: adamjernst

fbshipit-source-id: 9de91bb48e931c42b1042e163d47576c5422e2cc