Support smart invert colors for CKNetworkImageComponent
Summary:
In iOS 11, there's a new accessibility API to ignore the invert colors accessibility setting. This should be set for user generated content and icons.

For an image being downloaded over the network, there are almost no cases where we don't want to apply this setting.

If we really want, we can add an option to `CKNetworkImageComponentOptions` for an opt-out, but I don't feel this is necessary at this time.

Reviewed By: gkassabli

Differential Revision: D5587744

fbshipit-source-id: d845fa536f740eaa87c47c1da820ec5c8cd24799