Enable accessibility testing outside ComponentKit
Summary:
Right now, the ability to force accessibility to be enabled/disabled is only available within ComponentKit. This will expose the functionality publicly so that tools and tests outside of ComponentKit can use this as well.

Currently, once you call this, there's no way to go back to the default state of accessibility only being on when VoiceOver is on, so reset functionality was added as well.

Reviewed By: gkassabli

Differential Revision: D5829752

fbshipit-source-id: 19dd104c8daba946bfff41661b848f60de85b3cc