Fixes button sizing for multi-line labels
Summary:
`CKButtonComponent` used to use `-sizeWithFont:` to compute its size. `-sizeWithFont:` strips out `\n` before doing anything, so the size returned will always be the size of the text as if it were to be on a single line.

This causes issues with the following text: "Hello\nHow are you doing today my friend?" where `sizeWithFont:` would return a single long time (length of first line + length of the second) when really it needed to return the height of both lines and the width of the longest line. I've accessed the performance implications of such change and they seem to be trivial (and sometimes actually faster than the previous implementation) as long as sizes are computed on single line strings. Adding `\n` in your text will make its perfs degrade pretty quickly. Have a look at the linked task for a deeper perfs analysis.

Considering the current behavior doesn't account for multi-line strings, a slower approach only with multi-lines shouldn't have a negative effect of the overall performances if its unused.

I also got to remove the `#define tvOS` as `[NSString sizeWithAttributes:]` is available on tvOS.

Georgiy Kassabli, a few things:

- I started to add a `numberOfLines` attribute to `CKButtonComponentOptions` but then I realised that it will become probematic as it would default to `0` which isn't the current default of `1`. I then backtracked and removed the option. How do you think we should proceed on this? I don't recall being able to assign a default value to this if its not defined, which makes it complicated to differentiate the unlimited number of lines case and the default behavior of using 1. I _guess_ we could use a `NSNumber` instead of a plain `NSInteger` but that seems a bit yuck.

- I changed the default font size from `-[UIFont defaultButtonSize]` to `-[UIFont defaultLabelSize]` as it seems to be the actual default when no font is passed (and thus `button.titleLabel.font` is set to nil). Otherwise I think we'd have to keep `-[UIFont defaultButtonSize]` but default `nil` to `-[UIFont systemFontOfSize:[UIFont defaultButtonSize]]` in the attribute setter for `CKButtonComponent.titleFont`. Would that be preferred?

Thoughts?

Reviewed By: gkassabli

Differential Revision: D6008099

fbshipit-source-id: 946b65f5cf940f3ca9c79035ad21491da4b66179