Using separate memoizers for components and layouts
Summary:
So, I tried tackling the first limitation of the current memoization infra which is losing component cache when the same memoizer state is used for generating component tree followed by layout calculation. Component cache is not being carried over to a new memoizer state in that case, so subsequent component tree generation are not able to reuse anything.

I propose to fix this by having separate memoizers for components and layouts, so one does not affect the other.

The first thing I did is split `_CKComponentMemoizerImpl` into two separate classes — `CKComponentMemoizerState` and `CKComponentLayoutMemoizerState` that manage component cache and layout cache respectively. These classes also use different keys for TLS. There is some code duplication here right now, which we may want to eliminate later.

After that I made `CKComponentMemoizer` templated on the type of state it manages
(either component or layout cache), so you can only initialise it with a state of a matching type. The only thing that you need to do is to instantiate a memoizer with correct state, and it all works.

As a result, I was able to fix two tests:

- `testThatMemoizableComponentsAreMemoizedEvenWithLayoutCalled`
- `testComputeLayoutOnlyCalledOnceWhenEqualInputsAndWeSplitCreationAndLayout`

(*I think, these two were the tests I actually had in mind when I was asking about multiple tests testing the same thing.*)

Reviewed By: ocrickard

Differential Revision: D6147687

fbshipit-source-id: 78be9921bdcf4bed4d8d619e266d8734ffb86305