Fix pipeline based processing getting into an infinite loop
Summary:
We are seeing a lot of EXC/RESOURCE issues and Phil reported that he was seeing the data source sometimes gets into infinite loop. After looking at the code, I found that we are not correctly bailing out from the speculative execution.

This diff makes it so that 1) we only restart the async preparation from main thread when we fail to apply a computed modification 2) pass cancel state properly to recursive calls

Reviewed By: gkassabli

Differential Revision: D6297373

fbshipit-source-id: 85c49d1250e920f412755a7cd5f727de7a41ca5e