Using scope handles as keys in state update map
Summary:
[Discussion](https://fb.facebook.com/groups/782755281882662/permalink/883107718514084/)

We decided not to change scope handle identifiers themselves to encode parent / child relationships in scope tree, but instead maintain a link to the parent scope, and traverse this list when we need to determine whether a given scope is a parent of another scope. The problem with this approach is that at the actual point of use (for memoization, at least) we don't have actual scopes at hand, only scope handle identifiers, since that's what we store in state update map. This diff changes state update map to use scope handles as keys in the map

There are several kinds of changes in this diff that are needed to make this work:

- obviously, changing the declaration of `CKComponentStateUpdateMap` and `[CKComponentStateListener componentScopeHandle:rootIdentifier:didReceiveStateUpdate:userInfo:mode:]` to use actual handles instead of identifiers
- also quite obvious — changing the first parameter name from `globalIdentifier` to `handle`
- adding specialisation of `std::hash` for `CKComponentScopeHandle *`, which just uses the pointer as hash
- adding several imports of CKComponentScopeHandle.h, so the compiler can see the definition of `std::hash` specialisation (otherwise, it'll try to use the standard hash implementation, which tries to store an ARC-managed pointer to a scope handle in a union, which is a no-no)

Reviewed By: kfirapps

Differential Revision: D6339383

fbshipit-source-id: 7018cb771141b538a6a5460149b248053f894d26