Fixed memoization of scopeless components
Summary:
Turns out my previous solution wasn't handling scopeless components particularly well. If we tried to lookup a scopeless component in memoization cache, we still would check the **current scope** for pending state updates, which is actually a scope of a parent component, and that's not what we want here.

So I've changed the lookup logic to check whether component has a scope at all first, and only proceed to checking for pending state updates if a component actually has a scope.

I also fixed existing tests so that:

- they consistently use scope root from previous invocation of `CKBuildComponent()`
- `CKTestMemoizedComponent` actually supplies an identifier

Reviewed By: gkassabli

Differential Revision: D6413883

fbshipit-source-id: ce86fc9a9142abe1374c9b8dbb97ed7dd65b3527