Add function to unmount cell components on dealloc
Summary:
Add a function that will cause a particular component layout to be unmounted when a particular object is deallocated.

This is intended to be used when the owner of some component hierarchy is deallocated and we want to make sure the components it holds on to are deallocated as well, to avoid leaking memory.

Reviewed By: supercfanx

Differential Revision: D6483978

fbshipit-source-id: e80e8610c3f891f79f0bdd99ade88213cccb412a