Fixing issue with reusing zero-sized layout
Summary:
Another attempt to fix the caching issue that was suggested by Pasquale. The main idea is to treat final sizes that we get from Yoga as *exact* size specs, and reuse cached layout only if it was also measured with the exact size spec. The code for doing this was already present, so hopefully we will be able to remove some code as well.

So, in order to determine whether we can or cannot reuse a cached layout, we would perform 4 checks:

1) Is caching turned off explicitly in component options (`CKFlexboxComponentStyle.disableCachingToWorkAroundBug`)? If yes, recalculate.
2) Can we reuse a cached layout if we treat final size specs as `.atMost`? (I'm oversimplifying a bit here)
3) Can we reuse a cached layout if we treat final size specs as `.exactly`?
4) Is the final size zero? If yes, reuse.

New behaviour is to check only 3) and 4).

Switching between old and new behaviour is controlled by `CKFlexboxComponent._reuseOnlyExactSizeSpecs`. The old behaviour is implemented in `canReuseCachedLayout:forChildWithSize:`, the new behaviour is implemented in `canReuseCachedLayout:forChildWithExactSize:`. The former calls the latter, since it's a "superset" of checks.

For testing in prod, I'm planning to set this property via `CKComponentContext`, example of which you can see in the snapshot test. Unfortunately, `CKComponentContext` only allows ObjC objects as values, so I had to add a new class `CKFlexboxComponentContext` with the single property.

Reviewed By: gkassabli

Differential Revision: D6858075

fbshipit-source-id: 969ad94655c023c84f579876b8c7ea424bef83d2