Updated support for CKButtonComponent to size appropriately for number of lines
Summary:
CKButtonComponent recently added changes to allow the user to predfine the max number of lines for the button's titleLabel.
After doing some research, the NSString sizeWithAttributes API produces the absolute minimum bounding size for the text without considering
the numberOfLines property (because it's a UILabel property). The interesting part is that the produced bounding size height is actually
the same as the UIFont's lineHeight property. So producing a proper bounding size with a fixed number of lines is simple math.
boundingSize.height * numberOfLines.

Reviewed By: nscoding

Differential Revision: D6954627

fbshipit-source-id: 9927be0fdc14c2dbec521e9f56ad6cfa2c1364cd