Introduce 'didPrepareLayout:forComponent:' life cycle method
Summary:
Add a new life cycle to the component controller 'didPrepareLayout:forComponent:'.

This method should be called either from the data source or from an hosting view.

Data source:
This method should be called when a changeset was applied on the main thread for either insert or update.
In this stage there is a new component with a new computed layout.

Hosting View:
This method should be called when a new layout has been calculated and a new component has been attached to the component controller.

Reviewed By: gkassabli, Andrey-Mishanin

Differential Revision: D6425796

fbshipit-source-id: 19c00b5a93c7b6933e3b348cbd933491855eefca