Adding callbacks to CKDataSource triggered when new change has finished building
Summary:
CKDataSourceListener doesn't allow us to track a few events happening at the start of modification processing - start of change calculation and finish of it.
This diff adds optional callbacks that are triggered at those points.
Also this diff exposes userInfo property for DataSourceModification to be used in modification start processing callback - we can use it for identification purposes. UserInfo already was present in most of the DataSourceModification types, we just weren't exposing it explicitly in protocol

Reviewed By: kfirapps, Andrey-Mishanin

Differential Revision: D7111870

fbshipit-source-id: ebaea283e260a3b933eb8b532b831d0bc1328b3b