Adding CKDataSource listener method for synchronous modification appication
Summary:
We've previously added announcer callbacks that allow listener to track modification lifecycle. However one situation where we didn't get any callback until end of modification was synchronous application (either synchronously scheduled or forced with subsequent synchronous changeset).
This diff adds a callback that is called before modification in that case.
The reason we keep this callback separate from existing async callback for modification start is the scenario of forced sync application. In that scenario we might have some work done on work thread that will be thrown out and we want to track that as well.

Reviewed By: Andrey-Mishanin

Differential Revision: D7302402

fbshipit-source-id: 1abe789790deaa72f24657c78df212a042d9fd2f