Ensuring CKDataSource triggers controller invalidations on main thread when deallocated
Summary:
CKDataSource announce CKComponentController invalidation in deallocation. Previously this was done directly in dealloc, which could be triggered outside of MT. This caused invalidateController to be called on BG thread, thus breaking the contract that invalidateController in called only on Main Thread.
This diff makes sure we execute controller invalidation announcement only on Main Thread

Reviewed By: Andrey-Mishanin

Differential Revision: D7729164

fbshipit-source-id: b1c3aa0910941eae7da632870a0ae47cb8b4f733