Leak the global static mutex in CKTextKitContext to avoid static destructor fiasco
Summary:
OSS CK is crashing on the exit. The probable cause is that it is static mutex is destroyed at the same time when another thread tries to lock it.
This diff leaks the mutex in order to avoid that crash

I haven't tested this code, but this is what I would try to resolve https://github.com/facebook/componentkit/issues/892. Someone should run some smoke tests with something like the above
Closes https://github.com/facebook/componentkit/pull/906

Reviewed By: kfirapps

Differential Revision: D7616427

Pulled By: gkassabli

fbshipit-source-id: fe4d9fc1add8228d3bd672f853cca8d54a02bc58