Kill unnecessary CKButtonComponentAccessibilityConfiguration
Summary:
`CKButtonComponentAccessibilityConfiguration` was added in 2014, since then I believe the codebase has evolved a lot and we created a more cleaner struct `CKComponentAccessibilityContext` to replace the usage.

So to avoid more confusion moving forward, I modified a lot of files to get rid of `CKButtonComponentAccessibilityConfiguration`. We should all use `CKComponentAccessibilityContext` in the future.

Reviewed By: ikenwoo

Differential Revision: D7984088

fbshipit-source-id: d90017a8acac1ecee9a14c69dffc7eea11e1bc72