Scope tree diffing API
Summary:
The new `CK::animatedComponentsBetweenScopeRoots()` function takes two scope roots (the old and the new one) and returns a `CK::ComponentTreeDiff` containing components for which we'll need to run animations on initial mount (`appearedComponents`) and components for which we'll need to run change animations (`updatedComponents`). `updatedComponents` is actually a list of pairs containing the previous and the current generation of a particular "logical" component, so we can call `animationsFromPreviousComponent:` more easily.

The algorithm is fairly straightforward:

- Collect scope handles for all old and new components that have initial animations into 2 respective collections.
- The handles that are only in the second collection correspond to components that have just appeared and need their initial animations to be executed.
- Collect scope handles for all old and new components that have animations from previous component into 2 respective collections.
- The handles that are in both collections correspond to components that have been updated and need their animations from previous component to be executed.

Reviewed By: kfirapps

Differential Revision: D8146492

fbshipit-source-id: cc4b61ece4d5b5bb1abdc8c65d56967a8351d5ac