Introducing animation applicator for use at integration points
Summary:
`CK::AnimationApplicator` is a class template that maintains the state of animations infrastructure between mount passes.

At integration point you're supposed to pass an instance of `CKComponentAnimations` and a block that performs mounting and returns all components that were unmounted during that pass.

`CK::AnimationApplicator` will maintain a reference to a `CKComponentAnimationsController`, creating a new one if neccessary and asking the previous one to clean up animations for unmounted components.

Reviewed By: fabiomassimo

Differential Revision: D8624759

fbshipit-source-id: 9795eb42b63f0da69aeb7e2f208825f064c22681