Separate return type for CKComputeRootComponentLayout
Summary:
We may want to return additional information from the computation of the root layout so it makes sense to have a separate type for the result of this computation (not unlike `CKBuildComponentResult`).

Right now we have only one client for this functionality (component layout cache that currently is being stored in `extra` dictionary on the root layout) but there will be more in the future.

Reviewed By: kfirapps

Differential Revision: D8732191

fbshipit-source-id: 8e410e736b90262fe9b4ec612126a9c419ddfacf