Diffing layout trees instead of scope trees
Summary: Because the layout tree is what we are actually going to mount we need to diff layout trees to collect animations, not scope trees. Otherwise, we would consider every component with scope that was created, even if it doesn't get mounted eventually, and in that case we'd try to animate an unmounted component which is not allowed.

Reviewed By: aCorrado

Differential Revision: D8762676

fbshipit-source-id: 7ed00f711789974ebf42fdb4eb0e864bcdab9b88